+++ app/src/pages/dataset/DatasetPage.tsx
@@ -175,6 +175,7 @@ function DatasetPageContent({
     },
     [navigate, datasetId]
   );
+  const datasetHasVersions = (dataset.latestVersions?.edges.length ?? 0) > 0;
 
   // Set the initial tab
   const location = useLocation();
@@ -288,6 +289,7 @@ function DatasetPageContent({
             />
             <DatasetLabelConfigButton datasetId={dataset.id} />
             <Button
+              isDisabled={!datasetHasVersions}
               size="S"
               variant="primary"
               leadingVisual={<Icon svg={<Icons.PlayCircleOutline />} />}